fix: Payment allocation list currencies.
